+++ superbench/benchmarks/micro_benchmarks/__init__.py
@@ -11,7 +11,7 @@ from superbench.benchmarks.micro_benchmarks.computation_communication_overlap im
 from superbench.benchmarks.micro_benchmarks.cublas_function import CublasBenchmark
 from superbench.benchmarks.micro_benchmarks.blaslt_function_base import BlasLtBaseBenchmark
 from superbench.benchmarks.micro_benchmarks.cublaslt_function import CublasLtBenchmark
-from superbench.benchmarks.micro_benchmarks.hipblas_function import HipBlasLtBenchmark
+from superbench.benchmarks.micro_benchmarks.hipblaslt_function import HipBlasLtBenchmark
 from superbench.benchmarks.micro_benchmarks.cuda_gemm_flops_performance import CudaGemmFlopsBenchmark
 from superbench.benchmarks.micro_benchmarks.cuda_memory_bw_performance import CudaMemBwBenchmark
 from superbench.benchmarks.micro_benchmarks.cuda_nccl_bw_performance import CudaNcclBwBenchmark
